Move ClosureFiller to reusable module
