+++ crates/m3/src/builder/mod.rs
@@ -8,6 +8,7 @@ pub mod expr;
 mod multi_iter;
 pub mod statement;
 pub mod table;
+pub mod test_utils;
 pub mod types;
 pub mod witness;
