+++ crates/m3/src/gadgets/lookup.rs
@@ -82,49 +82,12 @@ impl LookupProducer {
 mod tests {
 	use std::{cmp::Reverse, iter, iter::repeat_with};
 
-	use binius_field::{arch::OptimalUnderlier128b, underlier::UnderlierType};
+	use binius_field::arch::OptimalUnderlier128b;
 	use bumpalo::Bump;
 	use rand::{rngs::StdRng, Rng, SeedableRng};
 
 	use super::*;
-	use crate::builder::{ConstraintSystem, Statement, TableFiller, TableId};
-
-	struct ClosureFiller<'a, U: UnderlierType, F: TowerField, Event> {
-		table_id: TableId,
-		fill: Box<
-			dyn for<'b> Fn(&'b [&'b Event], &'b mut TableWitnessSegment<U, F>) -> Result<()> + 'a,
-		>,
-	}
-
-	impl<'a, U: UnderlierType, F: TowerField, Event> ClosureFiller<'a, U, F, Event> {
-		fn new(
-			table_id: TableId,
-			fill: impl for<'b> Fn(&'b [&'b Event], &'b mut TableWitnessSegment<U, F>) -> Result<()> + 'a,
-		) -> Self {
-			Self {
-				table_id,
-				fill: Box::new(fill),
-			}
-		}
-	}
-
-	impl<'a, U: UnderlierType, F: TowerField, Event: Clone> TableFiller<U, F>
-		for ClosureFiller<'a, U, F, Event>
-	{
-		type Event = Event;
-
-		fn id(&self) -> TableId {
-			self.table_id
-		}
-
-		fn fill<'b>(
-			&'b self,
-			rows: impl Iterator<Item = &'b Self::Event> + Clone,
-			witness: &'b mut TableWitnessSegment<U, F>,
-		) -> Result<()> {
-			(*self.fill)(&rows.collect::<Vec<_>>(), witness)
-		}
-	}
+	use crate::builder::{test_utils::ClosureFiller, ConstraintSystem, Statement};
 
 	#[test]
 	fn test_basic_lookup_producer() {
